added C++ variant of replication catchup

+++ arangod/Replication/ContinuousSyncer.cpp
@@ -72,7 +72,9 @@ ContinuousSyncer::ContinuousSyncer(
       _verbose(configuration->_verbose),
       _masterIs27OrHigher(false),
       _hasWrittenState(false),
-      _supportsSingleOperations(false) {
+      _supportsSingleOperations(false),
+      _ignoreRenameCreateDrop(false),
+      _transientApplierState(false) {
   uint64_t c = configuration->_chunkSize;
   if (c == 0) {
     c = static_cast<uint64_t>(256 * 1024);  // 256 kb
@@ -102,11 +104,10 @@ ContinuousSyncer::~ContinuousSyncer() {
   abortOngoingTransactions(); 
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief run method, performs continuous synchronization
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::run() {
+  TRI_ASSERT(!_transientApplierState);
+
   if (_client == nullptr || _connection == nullptr || _endpoint == nullptr) {
     return TRI_ERROR_INTERNAL;
   }
@@ -280,10 +281,7 @@ retry:
   return TRI_ERROR_NO_ERROR;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief abort all ongoing transactions
-////////////////////////////////////////////////////////////////////////////////
-
 void ContinuousSyncer::abortOngoingTransactions() {
   try {
     // abort all running transactions
@@ -302,41 +300,33 @@ void ContinuousSyncer::abortOngoingTransactions() {
   }
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief set the applier progress
-////////////////////////////////////////////////////////////////////////////////
-
 void ContinuousSyncer::setProgress(char const* msg) {
+  if (_verbose) {
+    LOG_TOPIC(INFO, Logger::REPLICATION) << msg;
+  }
+  else {
+    LOG_TOPIC(DEBUG, Logger::REPLICATION) << msg;
+  }
+
+  if (_transientApplierState) {
+    return;
+  }
+
   _applier->setProgress(msg, true);
-
-  if (_verbose) {
-    LOG_TOPIC(INFO, Logger::REPLICATION) << msg;
-  }
-  else {
-    LOG_TOPIC(DEBUG, Logger::REPLICATION) << msg;
-  }
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief set the applier progress
-////////////////////////////////////////////////////////////////////////////////
-
 void ContinuousSyncer::setProgress(std::string const& msg) {
-  _applier->setProgress(msg.c_str(), true);
-  
-  if (_verbose) {
-    LOG_TOPIC(INFO, Logger::REPLICATION) << msg;
-  }
-  else {
-    LOG_TOPIC(DEBUG, Logger::REPLICATION) << msg;
-  }
+  setProgress(msg.c_str());
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief save the current applier state
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::saveApplierState() {
+  if (_transientApplierState) {
+    return TRI_ERROR_NO_ERROR;
+  }
+
   LOG_TOPIC(TRACE, Logger::REPLICATION) << "saving replication applier state. last applied continuous tick: " << _applier->_state._lastAppliedContinuousTick << ", safe resume tick: " << _applier->_state._safeResumeTick;
 
   int res = TRI_SaveStateReplicationApplier(_vocbase, &_applier->_state, false);
@@ -347,10 +337,7 @@ int ContinuousSyncer::saveApplierState() {
 
   return res;
 }
-////////////////////////////////////////////////////////////////////////////////
 /// @brief whether or not a marker should be skipped
-////////////////////////////////////////////////////////////////////////////////
-
 bool ContinuousSyncer::skipMarker(TRI_voc_tick_t firstRegularTick,
                                   VPackSlice const& slice) const {
   bool tooOld = false;
@@ -405,10 +392,7 @@ bool ContinuousSyncer::skipMarker(TRI_voc_tick_t firstRegularTick,
   return false;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief whether or not a collection should be excluded
-////////////////////////////////////////////////////////////////////////////////
-
 bool ContinuousSyncer::excludeCollection(std::string const& masterName) const {
   if (masterName[0] == '_' && !_includeSystem) {
     // system collection
@@ -433,11 +417,9 @@ bool ContinuousSyncer::excludeCollection(std::string const& masterName) const {
   return false;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief get local replication apply state
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::getLocalState(std::string& errorMsg) {
+  TRI_ASSERT(!_transientApplierState);
   uint64_t oldTotalRequests = _applier->_state._totalRequests;
   uint64_t oldTotalFailedConnects = _applier->_state._totalFailedConnects;
 
@@ -477,10 +459,7 @@ int ContinuousSyncer::getLocalState(std::string& errorMsg) {
   return res;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief inserts a document, based on the VelocyPack provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::processDocument(TRI_replication_operation_e type,
                                       VPackSlice const& slice,
                                       std::string& errorMsg) {
@@ -588,7 +567,7 @@ int ContinuousSyncer::processDocument(TRI_replication_operation_e type,
 
     Result res = trx.begin();
 
-    // fix error handling here when function returns result //////////////////////
+    // fix error handling here when function returns result    
     if (!res.ok()) {
       errorMsg = "unable to create replication transaction: " + res.errorMessage();
       res.reset(res.errorNumber(),errorMsg);
@@ -599,17 +578,13 @@ int ContinuousSyncer::processDocument(TRI_replication_operation_e type,
         res.reset();
       }
     }
-    // fix error handling here when function returns result //////////////////////
-
+    // fix error handling here when function returns result
     res = trx.finish(res);
     return res.errorNumber();
   }
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief starts a transaction, based on the VelocyPack provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::startTransaction(VPackSlice const& slice) {
   // {"type":2200,"tid":"230920705812199","collections":[{"cid":"230920700700391","operations":10}]}
 
@@ -653,10 +628,7 @@ int ContinuousSyncer::startTransaction(VPackSlice const& slice) {
   return res.errorNumber();
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief aborts a transaction, based on the VelocyPack provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::abortTransaction(VPackSlice const& slice) {
   // {"type":2201,"tid":"230920705812199","collections":[{"cid":"230920700700391","operations":10}]}
   std::string const id = VelocyPackHelper::getStringValue(slice, "tid", "");
@@ -694,10 +666,7 @@ int ContinuousSyncer::abortTransaction(VPackSlice const& slice) {
   return TRI_ERROR_REPLICATION_UNEXPECTED_TRANSACTION;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief commits a transaction, based on the VelocyPack provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::commitTransaction(VPackSlice const& slice) {
   // {"type":2201,"tid":"230920705812199","collections":[{"cid":"230920700700391","operations":10}]}
   std::string const id = VelocyPackHelper::getStringValue(slice, "tid", "");
@@ -735,10 +704,7 @@ int ContinuousSyncer::commitTransaction(VPackSlice const& slice) {
   return TRI_ERROR_REPLICATION_UNEXPECTED_TRANSACTION;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief renames a collection, based on the VelocyPack provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::renameCollection(VPackSlice const& slice) {
   if (!slice.isObject()) {
     return TRI_ERROR_REPLICATION_INVALID_RESPONSE;
@@ -770,11 +736,8 @@ int ContinuousSyncer::renameCollection(VPackSlice const& slice) {
   return _vocbase->renameCollection(col, name, true);
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief changes the properties of a collection, based on the VelocyPack
 /// provided
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::changeCollection(VPackSlice const& slice) {
   if (!slice.isObject()) {
     return TRI_ERROR_REPLICATION_INVALID_RESPONSE;
@@ -802,10 +765,7 @@ int ContinuousSyncer::changeCollection(VPackSlice const& slice) {
   return guard.collection()->updateProperties(data, doSync).errorNumber();
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief apply a single marker from the continuous log
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::applyLogMarker(VPackSlice const& slice,
                                      TRI_voc_tick_t firstRegularTick,
                                      std::string& errorMsg) {
@@ -852,6 +812,9 @@ int ContinuousSyncer::applyLogMarker(VPackSlice const& slice,
   }
 
   else if (type == REPLICATION_COLLECTION_CREATE) {
+    if (_ignoreRenameCreateDrop) {
+      return TRI_ERROR_NO_ERROR;
+    }
     if (slice.get("collection").isObject()) {
       return createCollection(slice.get("collection"), nullptr);
     }
@@ -859,10 +822,17 @@ int ContinuousSyncer::applyLogMarker(VPackSlice const& slice,
   }
 
   else if (type == REPLICATION_COLLECTION_DROP) {
+    if (_ignoreRenameCreateDrop) {
+      return TRI_ERROR_NO_ERROR;
+    }
     return dropCollection(slice, false);
   }
 
   else if (type == REPLICATION_COLLECTION_RENAME) {
+    if (_ignoreRenameCreateDrop) {
+      // do not execute rename operations
+      return TRI_ERROR_NO_ERROR;
+    }
     return renameCollection(slice);
   }
 
@@ -889,16 +859,17 @@ int ContinuousSyncer::applyLogMarker(VPackSlice const& slice,
   else if (type == REPLICATION_VIEW_CHANGE) {
     THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_NOT_IMPLEMENTED, "view change not yet implemented");
   }
+   
+  else if (_ignoreRenameCreateDrop) {
+    return TRI_ERROR_NO_ERROR;
+  }
 
   errorMsg = "unexpected marker type " + StringUtils::itoa(type);
 
   return TRI_ERROR_REPLICATION_UNEXPECTED_MARKER;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief apply the data from the continuous log
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::applyLog(SimpleHttpResult* response,
                                TRI_voc_tick_t firstRegularTick,
                                std::string& errorMsg,
@@ -1013,10 +984,7 @@ int ContinuousSyncer::applyLog(SimpleHttpResult* response,
   return TRI_ERROR_NO_ERROR;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief perform a continuous sync with the master
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::runContinuousSync(std::string& errorMsg) {
   static uint64_t const MinWaitTime = 300 * 1000;        // 0.30 seconds
   static uint64_t const MaxWaitTime = 60 * 1000 * 1000;  // 60 seconds
@@ -1184,10 +1152,7 @@ int ContinuousSyncer::runContinuousSync(std::string& errorMsg) {
   return TRI_ERROR_INTERNAL;
 }
 
-////////////////////////////////////////////////////////////////////////////////
 /// @brief fetch the initial master state
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::fetchMasterState(std::string& errorMsg,
                                        TRI_voc_tick_t fromTick,
                                        TRI_voc_tick_t toTick,
@@ -1311,11 +1276,8 @@ int ContinuousSyncer::fetchMasterState(std::string& errorMsg,
 
   return TRI_ERROR_NO_ERROR;
 }
-
-////////////////////////////////////////////////////////////////////////////////
+  
 /// @brief run the continuous synchronization
-////////////////////////////////////////////////////////////////////////////////
-
 int ContinuousSyncer::followMasterLog(std::string& errorMsg,
                                       TRI_voc_tick_t& fetchTick,
                                       TRI_voc_tick_t firstRegularTick,
@@ -1535,4 +1497,108 @@ int ContinuousSyncer::followMasterLog(std::string& errorMsg,
   }
 
   return TRI_ERROR_NO_ERROR;
+}
+
+/// @brief finalize the synchronization of a collection by tailing the WAL
+/// and filtering on the collection name until no more data is available
+int ContinuousSyncer::syncCollectionFinalize(std::string& errorMsg,
+                                             std::string const& collectionName,
+                                             TRI_voc_tick_t fetchTick) {
+  // fetch master state just once
+  int res = getMasterState(errorMsg);
+
+  if (res != TRI_ERROR_NO_ERROR) {
+    return res;
+  }
+
+  // we do not want to apply rename, create and drop collection operations
+  _ignoreRenameCreateDrop = true;
+  
+  // we do not want to store the applier status anywhere
+  _transientApplierState = true;
+
+  // TODO: _barrierId? Database name? _masterInfo!
+  
+  while (true) {
+    std::string const baseUrl = BaseUrl + 
+                                "/logger-follow?chunkSize=" + _chunkSize + 
+                        //        "&barrier=" + StringUtils::itoa(_barrierId) +
+                                "&from=" + StringUtils::itoa(fetchTick) +
+                                "&serverId=" + _localServerIdString + 
+                                "&collection=" + StringUtils::urlEncode(collectionName) +
+                                "&includeSystem=true";
+
+    // send request
+    std::string const progress =
+        "fetching master log for collection '" + collectionName + "' from tick " + StringUtils::itoa(fetchTick) +
+        ", barrier: " + StringUtils::itoa(_barrierId);
+    setProgress(progress);
+
+    std::unique_ptr<SimpleHttpResult> response(_client->request(rest::RequestType::GET, baseUrl, nullptr, 0));
+
+    if (response == nullptr || !response->isComplete()) {
+      errorMsg = "got invalid response from master at " +
+                 std::string(_masterInfo._endpoint) + ": " +
+                 _client->getErrorMessage();
+
+      return TRI_ERROR_REPLICATION_NO_RESPONSE;
+    }
+
+    TRI_ASSERT(response != nullptr);
+
+    if (response->wasHttpError()) {
+      errorMsg = "got invalid response from master at " +
+                 std::string(_masterInfo._endpoint) + ": HTTP " +
+                 StringUtils::itoa(response->getHttpReturnCode()) + ": " +
+                 response->getHttpReturnMessage();
+
+      return TRI_ERROR_REPLICATION_MASTER_ERROR;
+    }
+
+    bool checkMore = false;
+
+    bool found;
+    std::string header =
+        response->getHeaderField(TRI_REPLICATION_HEADER_CHECKMORE, found);
+
+    if (found) {
+      checkMore = StringUtils::boolean(header);
+    } else {
+      errorMsg = "got invalid response from master at " +
+                 std::string(_masterInfo._endpoint) +
+                 ": required header is missing";
+      return TRI_ERROR_REPLICATION_INVALID_RESPONSE;
+    }
+
+    header =
+        response->getHeaderField(TRI_REPLICATION_HEADER_LASTINCLUDED, found);
+    if (found) {
+      TRI_voc_tick_t lastIncludedTick = StringUtils::uint64(header);
+
+      if (lastIncludedTick > fetchTick) {
+        fetchTick = lastIncludedTick;
+      } else {
+        // we got the same tick again, this indicates we're at the end
+        checkMore = false;
+      }
+    } else {
+      errorMsg = "got invalid response from master at " +
+                 std::string(_masterInfo._endpoint) +
+                 ": required header is missing";
+      return TRI_ERROR_REPLICATION_INVALID_RESPONSE;
+    }
+
+    uint64_t processedMarkers = 0;
+    uint64_t ignoreCount = 0;
+    int res = applyLog(response.get(), fetchTick, errorMsg, processedMarkers, ignoreCount);
+
+    if (res != TRI_ERROR_NO_ERROR) {
+      return res;
+    }
+
+    if (!checkMore || processedMarkers == 0) {
+      // done!
+      return TRI_ERROR_NO_ERROR;
+    }
+  }
 }
